pyproject.toml
[tool.poetry]
name = "image_to_rgb"
version = "0.1.0"
description = ""
authors = ["ahmedMunna1767 <ahmedmunna172367@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
